+++ third-party/riscv-arch-tests/src/run-rv64i-arch-test.lua
@@ -41,10 +41,6 @@ local uarch_ram_image_filename = arg[1]
 local output_signature_file = arg[2]
 local uarch_ram_start = 0x70000000
 local uarch_ram_length = 0x1000000
-local dummy_rom_filename = os.tmpname()
-io.open(dummy_rom_filename, "w"):close()
-local deleter = {}
-setmetatable(deleter, { __gc = function() os.remove(dummy_rom_filename) end })
 
 local config = {
     uarch = {
@@ -54,7 +50,6 @@ local config = {
         },
     },
     processor = {},
-    rom = { image_filename = dummy_rom_filename },
     ram = { length = 0x1000 },
 }